Window partition and window reverse are now in SSR & uses L1 memory config Adds license comment in CAB __init__
