+++ models/experimental/SSR/tt/tile_refinement/RHAG/ATTEN_BLK/HAB/CAB/__init__.py
@@ -1,3 +1,6 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+# SPDX-License-Identifier: Apache-2.0
+
 from .CAB import TTCAB
 from .channel_attention import TTChannelAttention
